+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- updated `extension apply` command to support new request body schema
 - deprecate `deploy` command
 - update go version to 1.22.5